refactor(worktime-balance-chart): adapt integration test
